KoreanTTS/requirements.txt
# For MeloTTS (primary option)
melo-tts
librosa
soundfile
numpy

# For gTTS (fallback option)
gTTS